requirements.txt
asgiref==3.4.1
Django==3.2.10
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2-binary==2.8.6
pytz==2021.3
sqlparse==0.4.2
